requirements.txt
# Основные зависимости
openai-whisper==20250625
torch==2.10.0
torchaudio==2.10.0
pyTelegramBotAPI==4.31.0
moviepy==2.2.1
ffmpeg-python==0.2.0
numpy==2.4.2